refactor: reorganize project, RieDFMXxx naming, full type checking
- Move Python scripts from scripts/ to src/riedfm/cli/ module
  - scripts/ now contains bash wrappers only
  - Rename prepare_data → preprocess
  - Add console_scripts entry points in pyproject.toml
- Rename all classes to RieDFMXxx convention
  (ProductManifold → RieDFMProductManifold, REDFormer → RieDFMREDFormer, etc.)
- Add Makefile with targets: install, lint, format, test, pretrain, etc.
- Pre-commit: add mypy (full type checking), check-ast
- Fix all mypy type errors across the codebase
- Fix all ruff lint/format issues
- Add [tool.mypy] config with strict settings

+++ src/riedfm/utils/graph_metrics.py
@@ -58,8 +58,8 @@ def edge_type_accuracy(
         mask = true_types > 0
         if mask.sum() == 0:
             return 0.0
-        return (pred_types[mask] == true_types[mask]).float().mean().item()
+        return float((pred_types[mask] == true_types[mask]).float().mean().item())
     else:
         N = pred_types.shape[0]
         mask = ~torch.eye(N, dtype=torch.bool, device=pred_types.device)
-        return (pred_types[mask] == true_types[mask]).float().mean().item()
+        return float((pred_types[mask] == true_types[mask]).float().mean().item())
